Fix wrong Next link
The fixed line previously caused another check to fail, because it was comparing an integer with a float, e.g., `3 !== 3.0`. This could cause a "Next" link to appear on the last page of the Global List. With this line now returning an integer, the issue is gone.

+++ includes/image-sources/renderer/global-list.php
@@ -156,7 +156,7 @@ class Global_List extends Renderer {
 
 		// calculate total pages
 		if ( $per_page && $per_page < $total ) {
-			$up_limit = ceil( $total / $per_page );
+			$up_limit = (int) ceil( $total / $per_page );
 		}
 
 		ob_start();
